[131661] separate question controller for answer/comment
